==> 1_Construct_Dimensions.ipynb <==
%load_ext autoreload
%autoreload 2

from helpers.stateembed import *

models = read_models(path = 'models')

with open('data/wordlist/relationships.txt', 'r') as f:
    close_rel = f.readline().strip().split(', ')
    distant_rel = f.readline().strip().split(', ')

    close_rel = [x.strip() for x in close_rel]
    distant_rel = [x.strip() for x in distant_rel]

with open('data/wordlist/responsibility.txt', 'r') as f:
    high_resp = f.readline().strip().split(', ')
    low_resp = f.readline().strip().split(', ')

    high_resp = [x.strip() for x in high_resp]
    low_resp = [x.strip() for x in low_resp]

with open('data/wordlist/warmth.txt', 'r') as f:
    warm = f.readline().strip().split(', ')
    cold = f.readline().strip().split(', ')

    warm = [x.strip() for x in warm]
    cold = [x.strip() for x in cold]


with open('data/wordlist/supervisor.txt', 'r') as f:
    supervisors = f.readline().strip().split(', ')
    supervisors = [x.strip() for x in supervisors]

with open('data/wordlist/shame.txt', 'r') as f:
    shame = f.readline().strip().split(',')
    shame = [x.strip() for x in shame]

with open('data/wordlist/pride.txt', 'r') as f:
    pride = f.readline().strip().split(',')
    pride = [x.strip() for x in pride]

responsibility = Dimension(high_resp, low_resp)
relationships = Dimension(close_rel, distant_rel)
warmth = Dimension(warm, cold)

responsibilism = Concept(relationships, responsibility)
fuzzy_collectivism = Concept(Dimension(close_rel + distant_rel, []), warmth)

pd.Series(get_indicies_for_states(models, responsibilism), name='responsibilism') \
    .rename_axis('state') \
    .reset_index() \
    .to_csv('data/indices/responsibilism_df.csv', index=False)
    
pd.Series(get_indicies_for_states(models, fuzzy_collectivism), name='fuzzy_collectivism') \
    .rename_axis('state') \
    .reset_index() \
    .to_csv('data/indices/fuzzy_collectivism_df.csv', index=False)

### Theoretical Advances

responsibility_pride_shame = Dimension(pos = high_resp + shame, neg = low_resp + pride)
relationships_supervisors = Dimension(pos = close_rel + supervisors, neg = distant_rel)

responsibilism_v1 = Concept(relationships, responsibility_pride_shame)
responsibilism_v2 = Concept(relationships_supervisors, responsibility_pride_shame)
fuzzy_collectivism_v1 = Concept(relationships, warmth)

pd.Series(get_indicies_for_states(models, responsibilism_v1), name='responsibilism_v1') \
    .rename_axis('state') \
    .reset_index() \
    .to_csv('data/indices/responsibilism_v1_df.csv', index=False)

pd.Series(get_indicies_for_states(models, responsibilism_v2), name='responsibilism_v2') \
    .rename_axis('state') \
    .reset_index() \
    .to_csv('data/indices/responsibilism_v2_df.csv', index=False)

pd.Series(get_indicies_for_states(models, fuzzy_collectivism_v1), name='fuzzy_collectivism_v1') \
    .rename_axis('state') \
    .reset_index() \
    .to_csv('data/indices/fuzzy_collectivism_v1_df.csv', index=False)
