==> 2_Construct_Dataframe.ipynb <==
### Covariates

import pandas as pd

gallup = pd.read_excel('data/covariates/gallup_happiness.xlsx')
mitchell = pd.read_csv('data/covariates/mitchell.csv')
rocklage = pd.read_csv('data/covariates/rocklage_5.5.csv')
vader = pd.read_csv('data/covariates/vader.csv')

gallup.columns = gallup.columns.str.lower()
mitchell.columns = mitchell.columns.str.lower()
rocklage.columns = rocklage.columns.str.lower()
vader.columns = vader.columns.str.lower()

gallup['state'] = gallup['state'].str.lower().apply(lambda x: x.replace(' ', ''))
mitchell['state'] = mitchell['state'].str.lower().apply(lambda x: x.replace(' ', ''))
vader.rename(columns={'unnamed: 0': 'state'}, inplace=True)
rocklage.rename(columns={'filename': 'state'}, inplace=True)
rocklage['state'] = rocklage['state'].apply(lambda x: x.split('_')[0].replace(' ', ''))

new_rows = pd.DataFrame({
    'state': ['alabama', 'alaska', 'arkansas'],
    'valenceavg': [4.462381265958876, 5.224184259674828, 4.87671109519063]
})

rocklage = pd.concat([rocklage, new_rows], ignore_index=True)
rocklage.merge(gallup, on='state')[['well_being_index', 'valenceavg']].corr()

df = gallup.merge(mitchell, on='state') \
    .merge(vader, on='state') \
    .merge(rocklage, on='state')

df.loc[df['state'] == 'michigan', 'vandellocohencollectivismscore'] = 46

### Add Conceptual Indicies

fuzzy_collectivism = pd.read_csv('data/indices/fuzzy_collectivism_df.csv')
responsibilism = pd.read_csv('data/indices/responsibilism_df.csv')

responsibilism_v1 = pd.read_csv('data/indices/responsibilism_v1_df.csv')
responsibilism_v2 = pd.read_csv('data/indices/responsibilism_v2_df.csv')
fuzzy_collectivism_v1 = pd.read_csv('data/indices/fuzzy_collectivism_v1_df.csv')

df = df.merge(fuzzy_collectivism, on='state') \
    .merge(responsibilism, on='state') \
    .merge(responsibilism_v1, on='state') \
    .merge(responsibilism_v2, on='state') \
    .merge(fuzzy_collectivism_v1, on='state')

df.to_csv('data/df.csv', index=False)
