==> 3_Regression.ipynb <==
import pandas as pd

df = pd.read_csv('data/df.csv')

df[['vandellocohencollectivismscore', 'responsibilism']].corr()

df.sort_values('vandellocohencollectivismscore', ascending=False)[['state', 'responsibilism']]

df = df[df['state'] != 'nevada'].copy()

df[['state', 'responsibilism']]

df[['vandellocohencollectivismscore', 'valenceavg', 'responsibilism', 'fuzzy_collectivism', 'responsibilism_v1', 'responsibilism_v2', 'fuzzy_collectivism_v1']].corr()
